requirements.txt
schwab-py
python-dotenv
selenium